WIP: Placeholder for `Arch_kernelImagePaddr` unimplemented for riscv64
Likely to be called by `Arch_setKernelImage` (done via `kernelImageVPtr`
on arm32), which is also yet to be implemented for riscv64.

This changeset now builds and runs sel4test with multiple domains, as
there is no change to boot from mainline (haven't yet started invoking
map/clone methods to init kernel images for the different domains).

Signed-off-by: Robert Sison <robert.sison@unimelb.edu.au>